requirements.txt
hyperon
openai
python-dotenv
